catalog: Move builtin to catalog crate (#21791)
This commit moves the mz_adapter::catalog::builtin module to
mz_catalog::builtin which cleans up some duplicate structs.
