+++ src/catalog/Cargo.toml
@@ -9,19 +9,24 @@ publish = false
 [dependencies]
 futures = "0.3.25"
 itertools = "0.10.5"
+once_cell = "1.16.0"
 mz-audit-log = { path = "../audit-log" }
 mz-compute-client = { path = "../compute-client" }
 mz-controller = { path = "../controller" }
 mz-controller-types = { path = "../controller-types" }
 mz-ore = { path = "../ore", features = ["chrono", "async", "tracing_"] }
+mz-pgrepr = { path = "../pgrepr" }
 mz-proto = { path = "../proto" }
 mz-repr = { path = "../repr", features = ["tracing_"] }
 mz-sql = { path = "../sql" }
 mz-sql-parser = { path = "../sql-parser" }
 mz-stash = { path = "../stash" }
+mz-storage-client = { path = "../storage-client" }
 mz-storage-types = { path = "../storage-types" }
 proptest = { version = "1.0.0", default-features = false, features = ["std"] }
 proptest-derive = { version = "0.3.0", features = ["boxed_union"] }
+serde = "1.0.152"
+serde_json = "1.0.89"
 tracing = "0.1.37"
 workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
