+++ src/catalog/src/builtin.rs
@@ -48,9 +48,8 @@ use mz_storage_client::healthcheck::{
 use once_cell::sync::Lazy;
 use serde::Serialize;
 
-use crate::catalog::DEFAULT_CLUSTER_REPLICA_NAME;
-
 pub const BUILTIN_PREFIXES: &[&str] = &["mz_", "pg_", "external_"];
+const BUILTIN_CLUSTER_REPLICA_NAME: &str = "r1";
 
 #[derive(Debug)]
 pub enum Builtin<T: 'static + TypeReference> {
@@ -106,16 +105,6 @@ pub struct BuiltinLog {
     pub variant: LogVariant,
     pub name: &'static str,
     pub schema: &'static str,
-}
-
-impl From<&BuiltinLog> for mz_catalog::BuiltinLog {
-    fn from(log: &BuiltinLog) -> Self {
-        Self {
-            variant: log.variant.clone(),
-            name: log.name,
-            schema: log.schema,
-        }
-    }
 }
 
 #[derive(Hash, Debug)]
@@ -184,16 +173,6 @@ pub struct BuiltinRole {
     pub attributes: RoleAttributes,
 }
 
-impl From<&BuiltinRole> for mz_catalog::BuiltinRole {
-    fn from(role: &BuiltinRole) -> Self {
-        Self {
-            id: role.id,
-            name: role.name,
-            attributes: role.attributes.clone(),
-        }
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct BuiltinCluster {
     /// Name of the cluster.
@@ -203,30 +182,12 @@ pub struct BuiltinCluster {
     pub privileges: &'static [MzAclItem],
 }
 
-impl From<&BuiltinCluster> for mz_catalog::BuiltinCluster {
-    fn from(cluster: &BuiltinCluster) -> Self {
-        Self {
-            name: cluster.name,
-            privileges: cluster.privileges,
-        }
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct BuiltinClusterReplica {
     /// Name of the compute replica.
     pub name: &'static str,
     /// Name of the cluster that this replica belongs to.
     pub cluster_name: &'static str,
-}
-
-impl From<&BuiltinClusterReplica> for mz_catalog::BuiltinClusterReplica {
-    fn from(replica: &BuiltinClusterReplica) -> Self {
-        Self {
-            name: replica.name,
-            cluster_name: replica.cluster_name,
-        }
-    }
 }
 
 /// Uniquely identifies the definition of a builtin object.
@@ -4953,7 +4914,7 @@ pub const MZ_SYSTEM_CLUSTER: BuiltinCluster = BuiltinCluster {
 };
 
 pub const MZ_SYSTEM_CLUSTER_REPLICA: BuiltinClusterReplica = BuiltinClusterReplica {
-    name: DEFAULT_CLUSTER_REPLICA_NAME,
+    name: BUILTIN_CLUSTER_REPLICA_NAME,
     cluster_name: MZ_SYSTEM_CLUSTER.name,
 };
 
@@ -4975,7 +4936,7 @@ pub const MZ_INTROSPECTION_CLUSTER: BuiltinCluster = BuiltinCluster {
 };
 
 pub const MZ_INTROSPECTION_CLUSTER_REPLICA: BuiltinClusterReplica = BuiltinClusterReplica {
-    name: DEFAULT_CLUSTER_REPLICA_NAME,
+    name: BUILTIN_CLUSTER_REPLICA_NAME,
     cluster_name: MZ_INTROSPECTION_CLUSTER.name,
 };
 
@@ -5368,730 +5329,5 @@ pub mod BUILTINS {
 
     pub fn iter() -> impl Iterator<Item = &'static Builtin<NameReference>> {
         BUILTINS_STATIC.iter()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::{BTreeMap, BTreeSet};
-    use std::env;
-    use std::time::{Duration, Instant};
-
-    use mz_expr::MirScalarExpr;
-    use mz_ore::now::{to_datetime, NOW_ZERO, SYSTEM_TIME};
-    use mz_ore::task;
-    use mz_pgrepr::oid::{FIRST_MATERIALIZE_OID, FIRST_UNPINNED_OID};
-    use mz_repr::{Datum, RowArena, Timestamp};
-    use mz_sql::catalog::{CatalogSchema, SessionCatalog};
-    use mz_sql::func::{Func, OP_IMPLS};
-    use mz_sql::names::{PartialItemName, ResolvedDatabaseSpecifier};
-    use mz_sql::plan::{
-        CoercibleScalarExpr, ExprContext, HirScalarExpr, PlanContext, QueryContext, QueryLifetime,
-        Scope, StatementContext,
-    };
-    use tokio_postgres::types::Type;
-    use tokio_postgres::NoTls;
-
-    use crate::catalog::{Catalog, CatalogItem, SYSTEM_CONN_ID};
-    use crate::coord::dataflows::{prep_scalar_expr, EvalTime, ExprPrepStyle};
-    use crate::session::Session;
-
-    use super::*;
-
-    // Connect to a running Postgres server and verify that our builtin
-    // types and functions match it, in addition to some other things.
-    #[mz_ore::test(tokio::test)]
-    #[cfg_attr(miri, ignore)] //  unsupported operation: can't call foreign function `TLS_client_method` on OS `linux`
-    async fn test_compare_builtins_postgres() {
-        async fn inner(catalog: Catalog) {
-            // Verify that all builtin functions:
-            // - have a unique OID
-            // - if they have a postgres counterpart (same oid) then they have matching name
-            let (client, connection) = tokio_postgres::connect(
-                &env::var("POSTGRES_URL").unwrap_or_else(|_| "host=localhost user=postgres".into()),
-                NoTls,
-            )
-            .await
-            .expect("failed to connect to Postgres");
-
-            task::spawn(|| "compare_builtin_postgres", async move {
-                if let Err(e) = connection.await {
-                    panic!("connection error: {}", e);
-                }
-            });
-
-            struct PgProc {
-                name: String,
-                arg_oids: Vec<u32>,
-                ret_oid: Option<u32>,
-                ret_set: bool,
-            }
-
-            struct PgType {
-                name: String,
-                ty: String,
-                elem: u32,
-                array: u32,
-                receive: Option<u32>,
-            }
-
-            struct PgOper {
-                oprresult: u32,
-                name: String,
-            }
-
-            let pg_proc: BTreeMap<_, _> = client
-                .query(
-                    "SELECT
-                    p.oid,
-                    proname,
-                    proargtypes,
-                    prorettype,
-                    proretset
-                FROM pg_proc p
-                JOIN pg_namespace n ON p.pronamespace = n.oid",
-                    &[],
-                )
-                .await
-                .expect("pg query failed")
-                .into_iter()
-                .map(|row| {
-                    let oid: u32 = row.get("oid");
-                    let pg_proc = PgProc {
-                        name: row.get("proname"),
-                        arg_oids: row.get("proargtypes"),
-                        ret_oid: row.get("prorettype"),
-                        ret_set: row.get("proretset"),
-                    };
-                    (oid, pg_proc)
-                })
-                .collect();
-
-            let pg_type: BTreeMap<_, _> = client
-                .query(
-                    "SELECT oid, typname, typtype::text, typelem, typarray, nullif(typreceive::oid, 0) as typreceive FROM pg_type",
-                    &[],
-                )
-                .await
-                .expect("pg query failed")
-                .into_iter()
-                .map(|row| {
-                    let oid: u32 = row.get("oid");
-                    let pg_type = PgType {
-                        name: row.get("typname"),
-                        ty: row.get("typtype"),
-                        elem: row.get("typelem"),
-                        array: row.get("typarray"),
-                        receive: row.get("typreceive"),
-                    };
-                    (oid, pg_type)
-                })
-                .collect();
-
-            let pg_oper: BTreeMap<_, _> = client
-                .query("SELECT oid, oprname, oprresult FROM pg_operator", &[])
-                .await
-                .expect("pg query failed")
-                .into_iter()
-                .map(|row| {
-                    let oid: u32 = row.get("oid");
-                    let pg_oper = PgOper {
-                        name: row.get("oprname"),
-                        oprresult: row.get("oprresult"),
-                    };
-                    (oid, pg_oper)
-                })
-                .collect();
-
-            let conn_catalog = catalog.for_system_session();
-            let resolve_type_oid = |item: &str| {
-                conn_catalog
-                    .resolve_item(&PartialItemName {
-                        database: None,
-                        // All functions we check exist in PG, so the types must, as
-                        // well
-                        schema: Some(PG_CATALOG_SCHEMA.into()),
-                        item: item.to_string(),
-                    })
-                    .expect("unable to resolve type")
-                    .oid()
-            };
-
-            let func_oids: BTreeSet<_> = BUILTINS::funcs()
-                .flat_map(|f| f.inner.func_impls().into_iter().map(|f| f.oid))
-                .collect();
-
-            let mut all_oids = BTreeSet::new();
-
-            // A function to determine if two oids are equivalent enough for these tests. We don't
-            // support some types, so map exceptions here.
-            let equivalent_types: BTreeSet<(Option<u32>, Option<u32>)> = BTreeSet::from_iter(
-                [
-                    // We don't support NAME.
-                    (Type::NAME, Type::TEXT),
-                    (Type::NAME_ARRAY, Type::TEXT_ARRAY),
-                    // We don't support time with time zone.
-                    (Type::TIME, Type::TIMETZ),
-                    (Type::TIME_ARRAY, Type::TIMETZ_ARRAY),
-                ]
-                .map(|(a, b)| (Some(a.oid()), Some(b.oid()))),
-            );
-            let ignore_return_types: BTreeSet<u32> = BTreeSet::from([
-                1619, // pg_typeof: TODO: We now have regtype and can correctly implement this.
-            ]);
-            let is_same_type = |fn_oid: u32, a: Option<u32>, b: Option<u32>| -> bool {
-                if ignore_return_types.contains(&fn_oid) {
-                    return true;
-                }
-                if equivalent_types.contains(&(a, b)) || equivalent_types.contains(&(b, a)) {
-                    return true;
-                }
-                a == b
-            };
-
-            for builtin in BUILTINS::iter() {
-                match builtin {
-                    Builtin::Type(ty) => {
-                        assert!(all_oids.insert(ty.oid), "{} reused oid {}", ty.name, ty.oid);
-
-                        if ty.oid >= FIRST_MATERIALIZE_OID {
-                            // High OIDs are reserved in Materialize and don't have
-                            // PostgreSQL counterparts.
-                            continue;
-                        }
-
-                        // For types that have a PostgreSQL counterpart, verify that
-                        // the name and oid match.
-                        let pg_ty = pg_type.get(&ty.oid).unwrap_or_else(|| {
-                            panic!("pg_proc missing type {}: oid {}", ty.name, ty.oid)
-                        });
-                        assert_eq!(
-                            ty.name, pg_ty.name,
-                            "oid {} has name {} in postgres; expected {}",
-                            ty.oid, pg_ty.name, ty.name,
-                        );
-
-                        assert_eq!(
-                            ty.details.typreceive_oid, pg_ty.receive,
-                            "type {} has typreceive OID {:?} in mz but {:?} in pg",
-                            ty.name, ty.details.typreceive_oid, pg_ty.receive,
-                        );
-
-                        if let Some(typreceive_oid) = ty.details.typreceive_oid {
-                            assert!(
-                                func_oids.contains(&typreceive_oid),
-                                "type {} has typreceive OID {} that does not exist in pg_proc",
-                                ty.name,
-                                typreceive_oid,
-                            );
-                        }
-
-                        // Ensure the type matches.
-                        match &ty.details.typ {
-                            CatalogType::Array { element_reference } => {
-                                let elem_ty = BUILTINS::iter()
-                                    .filter_map(|builtin| match builtin {
-                                        Builtin::Type(ty @ BuiltinType { name, .. })
-                                            if element_reference == name =>
-                                        {
-                                            Some(ty)
-                                        }
-                                        _ => None,
-                                    })
-                                    .next();
-                                let elem_ty = match elem_ty {
-                                    Some(ty) => ty,
-                                    None => {
-                                        panic!("{} is unexpectedly not a type", element_reference)
-                                    }
-                                };
-                                assert_eq!(
-                                    pg_ty.elem, elem_ty.oid,
-                                    "type {} has mismatched element OIDs",
-                                    ty.name
-                                )
-                            }
-                            CatalogType::Pseudo => {
-                                assert_eq!(
-                                    pg_ty.ty, "p",
-                                    "type {} is not a pseudo type as expected",
-                                    ty.name
-                                )
-                            }
-                            CatalogType::Range { .. } => {
-                                assert_eq!(
-                                    pg_ty.ty, "r",
-                                    "type {} is not a range type as expected",
-                                    ty.name
-                                );
-                            }
-                            _ => {
-                                assert_eq!(
-                                    pg_ty.ty, "b",
-                                    "type {} is not a base type as expected",
-                                    ty.name
-                                )
-                            }
-                        }
-
-                        // Ensure the array type reference is correct.
-                        let schema = catalog
-                            .resolve_schema_in_database(
-                                &ResolvedDatabaseSpecifier::Ambient,
-                                ty.schema,
-                                &SYSTEM_CONN_ID,
-                            )
-                            .expect("unable to resolve schema");
-                        let allocated_type = catalog
-                            .resolve_entry(
-                                None,
-                                &vec![(ResolvedDatabaseSpecifier::Ambient, schema.id().clone())],
-                                &PartialItemName {
-                                    database: None,
-                                    schema: Some(schema.name().schema.clone()),
-                                    item: ty.name.to_string(),
-                                },
-                                &SYSTEM_CONN_ID,
-                            )
-                            .expect("unable to resolve type");
-                        let ty = if let CatalogItem::Type(ty) = &allocated_type.item {
-                            ty
-                        } else {
-                            panic!("unexpectedly not a type")
-                        };
-                        match ty.details.array_id {
-                            Some(array_id) => {
-                                let array_ty = catalog.get_entry(&array_id);
-                                assert_eq!(
-                                    pg_ty.array, array_ty.oid,
-                                    "type {} has mismatched array OIDs",
-                                    allocated_type.name.item,
-                                );
-                            }
-                            None => assert_eq!(
-                                pg_ty.array, 0,
-                                "type {} does not have an array type in mz but does in pg",
-                                allocated_type.name.item,
-                            ),
-                        }
-                    }
-                    Builtin::Func(func) => {
-                        for imp in func.inner.func_impls() {
-                            assert!(
-                                all_oids.insert(imp.oid),
-                                "{} reused oid {}",
-                                func.name,
-                                imp.oid
-                            );
-
-                            // For functions that have a postgres counterpart, verify that the name and
-                            // oid match.
-                            let pg_fn = if imp.oid >= FIRST_UNPINNED_OID {
-                                continue;
-                            } else {
-                                pg_proc.get(&imp.oid).unwrap_or_else(|| {
-                                    panic!(
-                                        "pg_proc missing function {}: oid {}",
-                                        func.name, imp.oid
-                                    )
-                                })
-                            };
-                            assert_eq!(
-                                func.name, pg_fn.name,
-                                "funcs with oid {} don't match names: {} in mz, {} in pg",
-                                imp.oid, func.name, pg_fn.name
-                            );
-
-                            // Complain, but don't fail, if argument oids don't match.
-                            // TODO: make these match.
-                            let imp_arg_oids = imp
-                                .arg_typs
-                                .iter()
-                                .map(|item| resolve_type_oid(item))
-                                .collect::<Vec<_>>();
-
-                            if imp_arg_oids != pg_fn.arg_oids {
-                                println!(
-                                "funcs with oid {} ({}) don't match arguments: {:?} in mz, {:?} in pg",
-                                imp.oid, func.name, imp_arg_oids, pg_fn.arg_oids
-                            );
-                            }
-
-                            let imp_return_oid = imp.return_typ.map(|item| resolve_type_oid(item));
-
-                            assert!(
-                                is_same_type(imp.oid, imp_return_oid, pg_fn.ret_oid),
-                                "funcs with oid {} ({}) don't match return types: {:?} in mz, {:?} in pg",
-                                imp.oid, func.name, imp_return_oid, pg_fn.ret_oid
-                            );
-
-                            assert_eq!(
-                                imp.return_is_set,
-                                pg_fn.ret_set,
-                                "funcs with oid {} ({}) don't match set-returning value: {:?} in mz, {:?} in pg",
-                                imp.oid, func.name, imp.return_is_set, pg_fn.ret_set
-                            );
-                        }
-                    }
-                    _ => (),
-                }
-            }
-
-            for (op, func) in OP_IMPLS.iter() {
-                for imp in func.func_impls() {
-                    assert!(all_oids.insert(imp.oid), "{} reused oid {}", op, imp.oid);
-
-                    // For operators that have a postgres counterpart, verify that the name and oid match.
-                    let pg_op = if imp.oid >= FIRST_UNPINNED_OID {
-                        continue;
-                    } else {
-                        pg_oper.get(&imp.oid).unwrap_or_else(|| {
-                            panic!("pg_operator missing operator {}: oid {}", op, imp.oid)
-                        })
-                    };
-
-                    assert_eq!(*op, pg_op.name);
-
-                    let imp_return_oid = imp
-                        .return_typ
-                        .map(|item| resolve_type_oid(item))
-                        .expect("must have oid");
-                    if imp_return_oid != pg_op.oprresult {
-                        panic!(
-                            "operators with oid {} ({}) don't match return typs: {} in mz, {} in pg",
-                            imp.oid,
-                            op,
-                            imp_return_oid,
-                            pg_op.oprresult
-                        );
-                    }
-                }
-            }
-        }
-
-        Catalog::with_debug(NOW_ZERO.clone(), inner).await
-    }
-
-    // Execute all builtin functions with all combinations of arguments from interesting datums.
-    #[mz_ore::test(tokio::test)]
-    #[cfg_attr(miri, ignore)] //  unsupported operation: can't call foreign function `TLS_client_method` on OS `linux`
-    async fn test_smoketest_all_builtins() {
-        fn inner(catalog: Catalog) {
-            let conn_catalog = catalog.for_system_session();
-
-            let resolve_type_oid = |item: &str| {
-                conn_catalog
-                    .resolve_item(&PartialItemName {
-                        database: None,
-                        schema: Some(PG_CATALOG_SCHEMA.into()),
-                        item: item.to_string(),
-                    })
-                    .unwrap_or_else(|_| panic!("unable to resolve type: {item}"))
-                    .oid()
-            };
-            let pcx = PlanContext::zero();
-            let scx = StatementContext::new(Some(&pcx), &conn_catalog);
-            let qcx = QueryContext::root(&scx, QueryLifetime::OneShot);
-            let ecx = ExprContext {
-                qcx: &qcx,
-                name: "smoketest",
-                scope: &Scope::empty(),
-                relation_type: &RelationType::empty(),
-                allow_aggregates: false,
-                allow_subqueries: false,
-                allow_parameters: false,
-                allow_windows: false,
-            };
-            let arena = RowArena::new();
-            let mut session = Session::dummy();
-            session
-                .start_transaction(to_datetime(0), None, None)
-                .expect("must succeed");
-            let prep_style = ExprPrepStyle::OneShot {
-                logical_time: EvalTime::Time(Timestamp::MIN),
-                session: &session,
-            };
-            // Extracted during planning; always panics when executed.
-            let ignore_names = BTreeSet::from([
-                "avg",
-                "avg_internal_v1",
-                "bool_and",
-                "bool_or",
-                "mod",
-                "mz_panic",
-                "mz_sleep",
-                "pow",
-                "stddev_pop",
-                "stddev_samp",
-                "stddev",
-                "var_pop",
-                "var_samp",
-                "variance",
-            ]);
-
-            let fns = BUILTINS::funcs()
-                .map(|func| (&func.name, func.inner))
-                .chain(OP_IMPLS.iter());
-
-            for (name, func) in fns {
-                if ignore_names.contains(name) {
-                    continue;
-                }
-                let Func::Scalar(impls) = func else {
-                    continue;
-                };
-
-                'outer: for imp in impls {
-                    let details = imp.details();
-                    let mut styps = Vec::new();
-                    for item in details.arg_typs.iter() {
-                        let oid = resolve_type_oid(item);
-                        let Ok(pgtyp) = mz_pgrepr::Type::from_oid(oid) else {
-                            continue 'outer;
-                        };
-                        styps.push(ScalarType::try_from(&pgtyp).expect("must exist"));
-                    }
-                    let datums = styps
-                        .iter()
-                        .map(|styp| {
-                            let mut datums = vec![Datum::Null];
-                            datums.extend(styp.interesting_datums());
-                            datums
-                        })
-                        .collect::<Vec<_>>();
-                    // Skip nullary fns.
-                    if datums.is_empty() {
-                        continue;
-                    }
-
-                    let return_oid = details
-                        .return_typ
-                        .map(|item| resolve_type_oid(item))
-                        .expect("must exist");
-                    let return_styp = &mz_pgrepr::Type::from_oid(return_oid)
-                        .ok()
-                        .map(|typ| ScalarType::try_from(&typ).expect("must exist"));
-
-                    let mut idxs = vec![0; datums.len()];
-                    let mut args = Vec::with_capacity(idxs.len());
-                    while idxs[0] < datums[0].len() {
-                        args.clear();
-                        for i in 0..(datums.len()) {
-                            args.push(datums[i][idxs[i]]);
-                        }
-
-                        let op = &imp.op;
-                        let scalars = args
-                            .iter()
-                            .enumerate()
-                            .map(|(i, datum)| {
-                                CoercibleScalarExpr::Coerced(HirScalarExpr::literal(
-                                    datum.clone(),
-                                    styps[i].clone(),
-                                ))
-                            })
-                            .collect();
-
-                        let call_name = format!(
-                            "{name}({}) (oid: {})",
-                            args.iter()
-                                .map(|d| d.to_string())
-                                .collect::<Vec<_>>()
-                                .join(", "),
-                            imp.oid
-                        );
-
-                        // Execute the function as much as possible, ensuring no panics occur, but
-                        // otherwise ignoring eval errors. We also do various other checks.
-                        let start = Instant::now();
-                        let res = (op.0)(&ecx, scalars, &imp.params, vec![]);
-                        if let Ok(hir) = res {
-                            if let Ok(mut mir) = hir.lower_uncorrelated() {
-                                // Populate unmaterialized functions.
-                                prep_scalar_expr(&catalog.state, &mut mir, prep_style.clone())
-                                    .expect("must succeed");
-
-                                if let Ok(eval_result_datum) = mir.eval(&[], &arena) {
-                                    if let Some(return_styp) = return_styp {
-                                        let mir_typ = mir.typ(&[]);
-                                        // MIR type inference should be consistent with the type
-                                        // we get from the catalog.
-                                        assert_eq!(mir_typ.scalar_type, *return_styp);
-                                        // The following will check not just that the scalar type
-                                        // is ok, but also catches if the function returned a null
-                                        // but the MIR type inference said "non-nullable".
-                                        if !eval_result_datum.is_instance_of(&mir_typ) {
-                                            panic!("{call_name}: expected return type of {return_styp:?}, got {eval_result_datum}");
-                                        }
-                                        // Check the consistency of `introduces_nulls` and
-                                        // `propagates_nulls` with `MirScalarExpr::typ`.
-                                        if let Some((introduces_nulls, propagates_nulls)) =
-                                            call_introduces_propagates_nulls(&mir)
-                                        {
-                                            if introduces_nulls {
-                                                // If the function introduces_nulls, then the return
-                                                // type should always be nullable, regardless of
-                                                // the nullability of the input types.
-                                                assert!(mir_typ.nullable, "fn named `{}` called on args `{:?}` (lowered to `{}`) yielded mir_typ.nullable: {}", name, args, mir, mir_typ.nullable);
-                                            } else {
-                                                let any_input_null =
-                                                    args.iter().any(|arg| arg.is_null());
-                                                if !any_input_null {
-                                                    assert!(!mir_typ.nullable, "fn named `{}` called on args `{:?}` (lowered to `{}`) yielded mir_typ.nullable: {}", name, args, mir, mir_typ.nullable);
-                                                } else {
-                                                    assert_eq!(mir_typ.nullable, propagates_nulls, "fn named `{}` called on args `{:?}` (lowered to `{}`) yielded mir_typ.nullable: {}", name, args, mir, mir_typ.nullable);
-                                                }
-                                            }
-                                        }
-                                        // Check that `MirScalarExpr::reduce` yields the same result
-                                        // as the real evaluation.
-                                        let mut reduced = mir.clone();
-                                        reduced.reduce(&[]);
-                                        match reduced {
-                                            MirScalarExpr::Literal(reduce_result, ctyp) => {
-                                                match reduce_result {
-                                                    Ok(reduce_result_row) => {
-                                                        let reduce_result_datum = reduce_result_row.unpack_first();
-                                                        assert_eq!(reduce_result_datum, eval_result_datum, "eval/reduce datum mismatch: fn named `{}` called on args `{:?}` (lowered to `{}`) evaluated to `{}` with typ `{:?}`, but reduced to `{}` with typ `{:?}`", name, args, mir, eval_result_datum, mir_typ.scalar_type, reduce_result_datum, ctyp.scalar_type);
-                                                        // Let's check that the types also match.
-                                                        // (We are not checking nullability here,
-                                                        // because it's ok when we know a more
-                                                        // precise nullability after actually
-                                                        // evaluating a function than before.)
-                                                        assert_eq!(ctyp.scalar_type, mir_typ.scalar_type, "eval/reduce type mismatch: fn named `{}` called on args `{:?}` (lowered to `{}`) evaluated to `{}` with typ `{:?}`, but reduced to `{}` with typ `{:?}`", name, args, mir, eval_result_datum, mir_typ.scalar_type, reduce_result_datum, ctyp.scalar_type);
-                                                    },
-                                                    Err(..) => {}, // It's ok, we might have given invalid args to the function
-                                                }
-                                            },
-                                            _ => unreachable!("all args are literals, so should have reduced to a literal"),
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        let elapsed = start.elapsed();
-                        if elapsed > Duration::from_millis(250) {
-                            panic!("LONG EXECUTION ({elapsed:?}): {call_name}");
-                        }
-
-                        // Advance to the next datum combination.
-                        for i in (0..datums.len()).rev() {
-                            idxs[i] += 1;
-                            if idxs[i] >= datums[i].len() {
-                                if i == 0 {
-                                    break;
-                                }
-                                idxs[i] = 0;
-                                continue;
-                            } else {
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        Catalog::with_debug(NOW_ZERO.clone(), |catalog| async { inner(catalog) }).await
-    }
-
-    /// If the given MirScalarExpr
-    ///  - is a function call, and
-    ///  - all arguments are literals
-    /// then it returns whether the called function (introduces_nulls, propagates_nulls).
-    fn call_introduces_propagates_nulls(mir_func_call: &MirScalarExpr) -> Option<(bool, bool)> {
-        match mir_func_call {
-            MirScalarExpr::CallUnary { func, expr } => {
-                if expr.is_literal() {
-                    Some((func.introduces_nulls(), func.propagates_nulls()))
-                } else {
-                    None
-                }
-            }
-            MirScalarExpr::CallBinary { func, expr1, expr2 } => {
-                if expr1.is_literal() && expr2.is_literal() {
-                    Some((func.introduces_nulls(), func.propagates_nulls()))
-                } else {
-                    None
-                }
-            }
-            MirScalarExpr::CallVariadic { func, exprs } => {
-                if exprs.iter().all(|arg| arg.is_literal()) {
-                    Some((func.introduces_nulls(), func.propagates_nulls()))
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    }
-
-    // Make sure pg views don't use types that only exist in Materialize.
-    #[mz_ore::test(tokio::test)]
-    #[cfg_attr(miri, ignore)] //  unsupported operation: can't call foreign function `TLS_client_method` on OS `linux`
-    async fn test_pg_views_forbidden_types() {
-        Catalog::with_debug(SYSTEM_TIME.clone(), |catalog| async move {
-            let conn_catalog = catalog.for_system_session();
-
-            for view in BUILTINS::views().filter(|view| {
-                view.schema == PG_CATALOG_SCHEMA || view.schema == INFORMATION_SCHEMA
-            }) {
-                let item = conn_catalog
-                    .resolve_item(&PartialItemName {
-                        database: None,
-                        schema: Some(view.schema.to_string()),
-                        item: view.name.to_string(),
-                    })
-                    .expect("unable to resolve view");
-                let full_name = conn_catalog.resolve_full_name(item.name());
-                for col_type in item
-                    .desc(&full_name)
-                    .expect("invalid item type")
-                    .iter_types()
-                {
-                    match &col_type.scalar_type {
-                        typ @ ScalarType::UInt16
-                        | typ @ ScalarType::UInt32
-                        | typ @ ScalarType::UInt64
-                        | typ @ ScalarType::MzTimestamp
-                        | typ @ ScalarType::List { .. }
-                        | typ @ ScalarType::Map { .. }
-                        | typ @ ScalarType::MzAclItem => {
-                            panic!("{typ:?} type found in {full_name}");
-                        }
-                        ScalarType::AclItem
-                        | ScalarType::Bool
-                        | ScalarType::Int16
-                        | ScalarType::Int32
-                        | ScalarType::Int64
-                        | ScalarType::Float32
-                        | ScalarType::Float64
-                        | ScalarType::Numeric { .. }
-                        | ScalarType::Date
-                        | ScalarType::Time
-                        | ScalarType::Timestamp
-                        | ScalarType::TimestampTz
-                        | ScalarType::Interval
-                        | ScalarType::PgLegacyChar
-                        | ScalarType::Bytes
-                        | ScalarType::String
-                        | ScalarType::Char { .. }
-                        | ScalarType::VarChar { .. }
-                        | ScalarType::Jsonb
-                        | ScalarType::Uuid
-                        | ScalarType::Array(_)
-                        | ScalarType::Record { .. }
-                        | ScalarType::Oid
-                        | ScalarType::RegProc
-                        | ScalarType::RegType
-                        | ScalarType::RegClass
-                        | ScalarType::Int2Vector
-                        | ScalarType::Range { .. }
-                        | ScalarType::PgLegacyName => {}
-                    }
-                }
-            }
-        })
-        .await
     }
 }
